requirements.txt
fastapi==0.109.2
youtube-transcript-api==0.6.2
uvicorn==0.27.1
pydantic==2.6.1
google-generativeai==0.4.0
supabase==2.3.4
pyjwt==2.8.0
